server/requirements.txt
langchain-mcp-adapters
langchain==0.3.24 
langchain-google-genai==2.1.3 
langchain-groq 
langchain_community 
tavily-python 
streamlit 
langchain_community
langchain-openai
python-dotenv
PyGithub
mcp-use
fastembed